+++ src/main/java/com/sktech/academicportal/repository/StudentRepository.java
@@ -3,6 +3,6 @@ package com.sktech.academicportal.repository;
 import com.sktech.academicportal.entity.StudentEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface StudentRepository extends JpaRepository<StudentEntity, Long> {
+public interface StudentRepository extends JpaRepository<StudentEntity, Integer> {
 
 }
